+++ src/types/prefectureCheckboxe.d.ts
@@ -1,0 +1,2 @@
+type PrefectureCheckboxes = Prefecture & { selected: boolean }
+type PrefectureCheckboxesForm = { prefectures: PrefectureCheckboxes[] | [] }
